create recycler home + Search view
