+++ app/src/main/AndroidManifest.xml
@@ -14,13 +14,13 @@
         android:theme="@style/Theme.AppMKE"
         tools:targetApi="31">
         <activity
-            android:name=".DescriptionActivity"
+            android:name=".ui.DescriptionActivity"
             android:exported="false" />
         <activity
-            android:name=".HomeActivity"
+            android:name=".ui.MainActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.HomeActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
