+++ app/src/main/java/com/example/appmke/back/productsAdapter.kt
@@ -1,0 +1,10 @@
+package com.example.appmke.back
+
+import androidx.recyclerview.widget.RecyclerView
+
+class productsAdapter():RecyclerView.Adapter<>() {
+
+}
+class productHolder {
+
+}
