+++ app/src/main/java/com/example/appmke/ui/HomeActivity.kt
@@ -1,0 +1,14 @@
+package com.example.appmke.ui
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.example.appmke.databinding.ActivityHomeBinding
+
+class HomeActivity : AppCompatActivity() {
+    lateinit var binding : ActivityHomeBinding
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding= ActivityHomeBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+    }
+}
