+++ app/src/main/java/com/example/appmke/ui/MainActivity.kt
@@ -1,7 +1,8 @@
-package com.example.appmke
+package com.example.appmke.ui
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.example.appmke.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
