+++ app/src/main/res/layout/item_products.xml
@@ -1,0 +1,113 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content">
+
+    <ImageView
+        android:id="@+id/iv_image_home"
+        android:layout_width="120dp"
+        android:layout_height="120dp"
+        android:layout_margin="10dp"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <LinearLayout
+        android:id="@+id/ll_description_home"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:orientation="vertical"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toEndOf="@id/iv_image_home"
+        app:layout_constraintTop_toTopOf="parent">
+
+        <TextView
+            android:id="@+id/tv_description_home"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:maxLines="3"
+            android:textSize="18sp" />
+
+        <LinearLayout
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:orientation="horizontal">
+
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="@string/home_tv_dollar_sign"
+                android:textSize="25sp" />
+
+            <TextView
+                android:id="@+id/tv_price_home"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginStart="5dp"
+                android:textSize="25sp" />
+
+            <TextView
+                android:id="@+id/tv_offers_percent_home"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginStart="10dp"
+                android:textSize="15sp" />
+
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginStart="5dp"
+                android:text="@string/home_tv_offers_percent"
+                android:textSize="15sp" />
+        </LinearLayout>
+
+        <LinearLayout
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:orientation="horizontal">
+
+            <ImageView
+                android:id="@+id/tv_star_a_home"
+                android:layout_width="10dp"
+                android:layout_height="20dp" />
+
+            <ImageView
+                android:id="@+id/tv_star_b_home"
+                android:layout_width="10dp"
+                android:layout_height="20dp" />
+
+            <ImageView
+                android:id="@+id/tv_star_c_home"
+                android:layout_width="10dp"
+                android:layout_height="20dp" />
+
+            <ImageView
+                android:id="@+id/tv_star_d_home"
+                android:layout_width="10dp"
+                android:layout_height="20dp" />
+
+            <ImageView
+                android:id="@+id/tv_star_e_home"
+                android:layout_width="10dp"
+                android:layout_height="20dp" />
+
+            <ImageView
+                android:id="@+id/tv_star_f_home"
+                android:layout_width="10dp"
+                android:layout_height="20dp" />
+
+            <ImageView
+                android:id="@+id/tv_star_g_home"
+                android:layout_width="10dp"
+                android:layout_height="20dp" />
+
+            <ImageView
+                android:id="@+id/tv_star_h_home"
+                android:layout_width="10dp"
+                android:layout_height="20dp" />
+        </LinearLayout>
+    </LinearLayout>
+
+
+</androidx.constraintlayout.widget.ConstraintLayout>
